requirements.txt
torch
torchvision
mlflow
dvc
fastapi
uvicorn
pytest
requests
prometheus-fastapi-instrumentator
scikit-learn
pandas
numpy
pillow
python-dotenv
httpx
python-multipart
